Update Cypress configuration and test workflow
This commit updates the Cypress configuration file (cypress.config.ts) to use single quotes for imports and adds the 'reporter' option with the value 'mochawesome'. It also updates the test workflow (test.yml) to include the 'reporter-options' flag for the Cypress run command, specifying the report directory, overwrite option, and JSON output. Additionally, the package.json file is modified to include the 'cypress-multi-reporters' and 'mochawesome' dependencies.